I've added a new class called ImageMenu.kt that reuses the layout of the main menu. It allows another menu to be called when the user presses into a certain category. For example, when the user clicks on 'Capital Cities' it will take them to another image menu where they can pick what continent they want to focus on.
Some of the images on the main menu had to be changed to pngs so the transparent background would work well for both light and dark mode. I've created a new kotlin class called Typing Quiz. The functionality isn't completed yet but the goal for this class is that it will be a different type of quiz where the user has to type in their answer instead of having multiple choice buttons.